1) refactoring controller/configuration classes 2) removing duplicate PrinterPrinter 3) moving AppProps into separate class
