+++ src/main/java/ru/otus/homework/MessageSourceConfig.java
@@ -1,0 +1,19 @@
+package ru.otus.homework;
+
+import org.springframework.context.MessageSource;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.ResourceBundleMessageSource;
+import org.springframework.stereotype.Component;
+
+@Configuration
+@Component
+public class MessageSourceConfig {
+    @Bean
+    public MessageSource messageSource() {
+        ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
+        messageSource.setBasenames("locale/message");
+        messageSource.setDefaultEncoding("UTF-8");
+        return messageSource;
+    }
+}
