+++ src/main/java/ru/otus/homework/beans/finalResultsPrinter/FinalResultsPrinter.java
@@ -1,4 +1,4 @@
-package ru.otus.homework.beans.finalResults;
+package ru.otus.homework.beans.finalResultsPrinter;
 
 import ru.otus.homework.model.Result;
 import ru.otus.homework.model.Student;
