+++ src/main/java/ru/otus/homework/beans/localeService/LocaleServiceImpl.java
@@ -1,16 +1,13 @@
 package ru.otus.homework.beans.localeService;
 
-import lombok.Data;
 import lombok.Setter;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import ru.otus.homework.beans.messageManager.MessageManagerImpl;
 import ru.otus.homework.exceptions.WrongInputsError;
+import ru.otus.homework.beans.AppProps;
 
 import java.util.Locale;
-import java.util.Map;
 import java.util.Scanner;
 
 @Service
@@ -19,14 +16,14 @@ public class LocaleServiceImpl implements LocaleService {
     private static final int MAX_ERROR_ATTEMPTS = 3;
     private int errorCounts = 0;
 
-    @Autowired
-    private AppProps appProps;
+    private final AppProps appProps;
 
     private final MessageManagerImpl messageManager;
 
     @Autowired
-    public LocaleServiceImpl(MessageManagerImpl messageManager) {
+    public LocaleServiceImpl(MessageManagerImpl messageManager, AppProps appProps) {
         this.messageManager = messageManager;
+        this.appProps = appProps;
     }
 
     public void setLocalization() throws WrongInputsError {
@@ -54,15 +51,5 @@ public class LocaleServiceImpl implements LocaleService {
             errorCounts++;
             setLocalization();
         }
-    }
-
-    @ConfigurationProperties(prefix = "lang")
-    @Component
-    @Data
-    public class AppProps {
-        //@Qualifier("allLangs")
-        private String allLangs;
-        //@Qualifier("lang_localization")
-        private Map<String, String> langLocalization;
     }
 }
